DTR-1322:Move packages according to PR comments
