+++ conf/app.routes
@@ -25,4 +25,4 @@ POST       /filing/delete/return-agent                              uk.gov.hmrc.
 
 POST       /filing/update/return-version                            uk.gov.hmrc.formpproxy.sdlt.controllers.returns.ReturnsController.updateReturnVersion()
 
-POST       /create/predefined-agent                                 uk.gov.hmrc.formpproxy.sdlt.controllers.manageAgents.CreatePredefinedAgentController.createPredefinedAgent()
+POST       /create/predefined-agent                                 uk.gov.hmrc.formpproxy.sdlt.controllers.agents.CreatePredefinedAgentController.createPredefinedAgent()
